Data.IntList: added and removed
Ignore-this: 671af833042641f7d4da18543b6b750e

This datastructure was an attempt at reducing memory overhead and
access latency for the memoized list of primes. However, it was not
successful in lowering memory requirements, so I've reverted back
to the original version. The file is left in situ, in case of future
attempts.

darcs-hash:5b2018239852ea66eda7b6bdf66bb6cb3640c91d